+++ huobi/requstclient.py
@@ -302,7 +302,7 @@ class RequestClient(object):
         The request of get open orders.
 
         :param symbol: The symbol, like "btcusdt". (mandatory)
-        :param account_type: The order side, buy or sell. If no side defined, will return all open orders of the account. (mandatory)
+        :param account_type: account type, all defination to see AccountType in SDK. (mandatory)
         :param side: The order side, buy or sell. If no side defined, will return all open orders of the account. (optional)
         :param size: The number of orders to return. Range is [1, 500]. Default is 100. (optional)
         :param direct: 1:prev  order by ID asc from from_id, 2:next order by ID desc from from_id